Switch expected to also be a double for sgd loss functions.
(This is in preparation for supporting regression in addition to
classification.)

+++ include/classify/loss/hinge.h
@@ -35,8 +35,8 @@ struct hinge : public loss_function
      */
     const static std::string id;
 
-    double loss(double prediction, int expected) const override;
-    double derivative(double prediction, int expected) const override;
+    double loss(double prediction, double expected) const override;
+    double derivative(double prediction, double expected) const override;
 };
 }
 }
