+++ src/classify/loss/perceptron.cpp
@@ -14,14 +14,14 @@ namespace loss
 
 const std::string perceptron::id = "perceptron";
 
-double perceptron::loss(double prediction, int expected) const
+double perceptron::loss(double prediction, double expected) const
 {
     if (prediction * expected <= 0)
         return -expected * prediction;
     return 0;
 }
 
-double perceptron::derivative(double prediction, int expected) const
+double perceptron::derivative(double prediction, double expected) const
 {
     if (prediction * expected <= 0)
         return -expected;
